requirements.txt
Flask==0.10.1
Flask-Admin==1.0.6
Flask-Cache==0.12
Flask-DebugToolbar==0.8.0
Flask-Login==0.2.7
Flask-OpenID==1.1.1
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.2
Jinja2==2.7.1
MarkupSafe==0.18
MySQL-python==1.2.4
SQLAlchemy==0.8.2
WTForms==1.0.5
Werkzeug==0.9.4
argparse==1.2.1
blinker==1.3
itsdangerous==0.24
pylibmc==1.3.0
python-openid==2.2.5
requests==2.2.1
steamodd==4.10
wsgiref==0.1.2
